+++ qord/core/rest.py
@@ -315,8 +315,17 @@ class RestClient:
         data = await self.request(route)
         return data
 
-    async def list_guild_members(self, guild_id: int, params: typing.Dict[str, typing.Any]):
+    async def get_guild_members(self, guild_id: int, limit: int = UNDEFINED, after: int = UNDEFINED):
         route = Route("GET", "/guilds/{guild_id}/members", guild_id=guild_id)
+
+        params = {}
+
+        if limit is not UNDEFINED:
+            params["limit"] = limit
+
+        if after is not UNDEFINED:
+            params["after"] = after
+
         data = await self.request(route, params=params)
         return data